Move 'getPerkDescriptions' out into a separate file and rename it to 'getPlugDescriptions'.
